New: checking connection on startup
<iio2> waits for an internet connection during startup <hook> and if it
succeeds, it tells the NTPDAEMON <app> that it should go online.
iio2 is python2 (for qtile is), relies on urllib2 and waits for
the reference webserver <app.REFERENCE> to answer.
While at it, constants in <app> are now pep8 conformist CAPITALIZED as I
strive for some consistency.
TODO: exploit the connection checker

peripherals/conkyrc: corrected typos